TST: isolate test_tools_vector unit tests
